+++ examples/simple/igraph_erdos_renyi_game_gnm.c
@@ -11,7 +11,7 @@ int main(void) {
 
     igraph_erdos_renyi_game_gnm(
         &graph, /* n= */ 100, /* m= */ 100,
-        IGRAPH_UNDIRECTED, IGRAPH_NO_LOOPS
+        IGRAPH_UNDIRECTED, IGRAPH_NO_LOOPS, IGRAPH_NO_MULTIPLE
     );
 
     /* Compute the fraction of vertices contained within the largest connected component */